fix: problem with prerendering of generating mnemonic
